+++ src/routes/graphql/resolvers/getUserByIdWithEntities/resolver.ts
@@ -1,5 +1,5 @@
 import { FastifyInstance } from "fastify/types/instance";
-import { MercuriusLoaders } from "mercurius";
+// import { MercuriusLoaders } from "mercurius";
 
 export const getUserByIdWithEntitiesResolver = {
   Query: {
@@ -39,17 +39,50 @@ export const getUserByIdWithEntitiesResolver = {
   // },
 };
 
-export const loaders: MercuriusLoaders = {
-  User: {
-    async profile(queries, ctx) {
-      return queries.map(({ obj, params }) => {
-        const { db } = ctx.app;
+// export const getUserByIdWithEntitiesLoaders: MercuriusLoaders = {
+//   UserWithEntities: {
+//     async profile(queries, ctx) {
+//       const { db } = ctx.app;
 
-        return db.profiles.findOne({
-          key: "userId",
-          equals: obj.id,
-        });
-      });
-    },
-  },
-};
+//       return queries.map(({ obj, params }) => {
+//         return db.profiles.findOne({
+//           key: "userId",
+//           equals: obj.id,
+//         });
+//       });
+//     },
+//     async posts(queries, ctx) {
+//       const { db } = ctx.app;
+
+//       return queries.map(async ({ obj, params }) => {
+//         const posts = await db.posts.findMany({
+//           key: "userId",
+//           equals: obj.id,
+//         });
+
+//         return posts.map((post) =>
+//           db.posts.findOne({
+//             key: "id",
+//             equals: post.id,
+//           })
+//         );
+//       });
+//     },
+//     async memberType(queries, ctx) {
+//       const { db } = ctx.app;
+
+//       return queries.map(async ({ obj, params }) => {
+//         const profile = await db.profiles.findOne({
+//           key: "userId",
+//           equals: obj.id,
+//         });
+
+//         if (profile)
+//           return db.memberTypes.findOne({
+//             key: "id",
+//             equals: profile.memberTypeId,
+//           });
+//       });
+//     },
+//   },
+// };
